+++ package.json
@@ -17,7 +17,9 @@
     "dayjs": "^1.11.13",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.511.0",
+    "moment": "^2.30.1",
     "next": "15.3.2",
+    "rc-time-picker": "^3.7.3",
     "react": "^19.0.0",
     "react-day-picker": "^9.7.0",
     "react-dom": "^19.0.0",
