+++ src/app/guest/profile/layout.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export default function GuestLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      <main>{children}</main>
+    </>
+  );
+}
